fixbug gia tien

+++ app/src/main/java/com/example/androidprojectbtl/Model/ProductModel.java
@@ -104,7 +104,7 @@ public class ProductModel implements Serializable, IProduct {
         this.hinhanh = hinhanh;
     }
 
-    public float getGiatien() {
+    public long getGiatien() {
         return giatien;
     }
